Add unit tests for Gas/Aerosol Exchange.
The unit tests are created by running the tests in mam_refactor
and printing the input/output arguments and hard-coding the
same values into the unit tests. The output of the tested
function is compared to the mam_refactor result.

Being a unit tests there is no Skywalker input file and the
function is only called a few dozen times but the test runs
very fast and does exercise the code.