Add unit test for gas_aerosol_uptake_rates_1box
There was difficulty in getting the regression test
working so backed off a bit and rearranged the code
to be a bit more like the Fortran and create this
function, gas_aerosol_uptake_rates_1box, to be just
like the Fortran version. Fixing the unit test found
many problems in the setup.