Get first skywkr_gasaerexch_timestepping test running. (#87)
* Get first skywkr_gasaerexch_timestepping test running.

When comparing the Skywalker output files by had, the test
compares with the mam_refactor Skywalker output to
about 6 digits of accuracy.  This seems like too large
a diff but could be due to differing universal constants
or some other non-logic difference. It is close enough
for now to commit and then explore differences after
the new year.

* Clang format.

* Cuda fixes.

* Code review changes.