Create a combined process test.
Based on the amicphys source in mam_refactor.

This calls most of the processes we now have in mam4xx
to pipe the output from one to the in put of the next.
Currently it just runs with the input taken from an
example in mam_refactor but does not check that the
output is correct. But it compiles and runs without
errors.