Adds a `make memcheck` target that runs tests with Valgrind.
I considered having our CI tests run with Valgrind/memcheck, but some of
the cross-validation tests take a very long time with this. If running
tests with Valgrind is interesting to us, we might consider altering
some of the larger cross-validation tests, like
`run_vehkamaki2002_fig8`, which generates a contour plot of the
nucleation rate as a function of 2 variables.