Add a test based on the amicphys source in mam_refactor.
This calls most of the processes we now have in
mam4xx to pass the output from one to the in put of the next.
So far only the clear sky functions check out and pass with
a relative epsilon of .001 This is above machine precision
but I think it is due to the changes in constants between
mam_refactor and mam4xx. For instance in mam4xx the some
properties are taken as a double precision log value while
in mam_refactor the values are hard coded as single precision
constants. Note that the jroverf/ChangesToMatch_mam4xx branch
of mam_refactor was used to fix a couple of the values
with constants to change mam_refactor to match mam4xx.