Adds NUM_VERTICAL_LEVELS as a CMake parameter that defines mam4::nlev.
Because of the way we use automatic C++ arrays in some of our
parameterizations, it is necessary for us to have the number of vertical
levels as a compile-time constant. While this is undesirable in an ideal
situation, it's actually fine in the short term because EAMxx also has a
compile-time-configured number of vertical levels.

mam4::nlev now contains the number of vertical levels used internally
within mam4xx's C++ parameterizations.  I haven't changed any of the test
or validation code to set the number of vertical levels to mam4::nlev,
since that would imply that we have datasets for other numbers of
vertical levels. To make it possible for us to run validations tests in
situations where mam4::nlev != 72, I have added the constraint that
mam4::nlev >= 72. As long as we are zeroing our views before using
them, that should allow these tests to run and pass.

I've also adjusted our CI workflow scripts to explicitly set the number
of vertical levels to 72. Strictly speaking, this isn't necessary,
because 72 is the default value for this parameter. But I think it helps
to have these things made explicit.

Closes #223