Adding a Sources container for storing surface and volume sources.
For now, I'm just adding the container. When it's appropriate, we can
add sources to the appropriate interfaces.  On the mam4xx side, we can
assume that e.g. emissions data are placed properly within Sources.
On the EAMxx side, we must make sure that such source data are properly
interpolated from files, etc.