Refactor the main loops within Wet Deposition. (#291)
* Refactor the main loops within Wet Deposition.

Wet deposition has loops within loops within loops
with column levels within species within phase
within modes. Previously I had tried to put the outer loop
to be the levels as this was the approach at the time.
This resulted in awkward code and restrictions
on the number of threads that could be run.

Now the inner loop is over levels with outer loops are over
modes, phases and species.  This requires a lot of parallel_for
loops on teams but since it is all within compute_tendencies
which is already on device, these are not a kernel launches
and should be just as fast as before.

Should have written it this way to begin with.

* Clamg format.