Trivial change to use aero_model::gas_pcnst.
The classes that set a gas_pcnst from aero_model::gas_pcnst
in public header data were changed to use the one from
aero_model directly. Clean up the code a bit.

Note that there is a gas_pcnst in gas_chemistry that is set
to 31. The one in aero_model is set to 40. Not sure why
there is a difference.