Fix where value is set.
I think this was a programming bug where a set statement inside
a loop should have been outside. Before it was getting set on
the last pass through the loop but on the first pass there was
a divide by zero. Moving the set statement outside the loop
fixes the divide by zero and does not change the result.