Fix for reordering of aerosols within modes.
Balwinder has a branch in which he changes the ordering
of aerosols within modes. This should not make a difference
but an extra call to aerosol_index_for_mode was breaking
Gas/Aerosol exchange in this case.  Once this fix is merged
into Balwinder's branch, it should fix his tests.