Fix indexing of factnum in ndrop.
It is a bit ironic given the comment in the code that
sub-views have to be carefully treated with respect to
the data() function as resulting 1D array might not be
continuous. This was the case for factnum. So created
an array of length 4 to use as the 1D array and copied
the strided sub-view into the continuous 1D array.